[ush_stm32f4xx_uart.h] feat: delete USH_USART_dmaPack enum

+++ TheTicker/Drivers/Custom/Inc/ush_dma.h
@@ -70,10 +70,10 @@ typedef struct
 	uint16_t FIFOThreshold;					/* This parameter specifies the FIFO threshold level.
 	 	 	 	 	 	 	 	 	 	 	   This parameter can be a value of @ref USH_DMA_fifoThresholdLevel */
 
-	uint8_t StreamBaseAddress;				/* DMA stream base address.
+	uint32_t StreamBaseAddress;				/* DMA stream base address.
 											   System parameter that is set using the DMA_calcBaseAndIndex function. */
 
-	uint8_t StreamIndex;					/* DMA stream index.
+	uint32_t StreamIndex;					/* DMA stream index.
 											   System parameter that is set using the DMA_calcBaseAndIndex function. */
 
 } USH_DMA_initTypeDef;
